accept prop instead of refuse
