+++ tests/RifmFormat.test.js
@@ -1,60 +1,7 @@
 // @flow
 
-import * as React from 'react';
-import TestRenderer, { act } from 'react-test-renderer';
-// import { act } from 'react-dom/test-utils';
-import { Value } from 'react-powerplug';
-import { Rifm } from '../src';
 import { numberFormat, currencyFormat, currencyFormat2 } from '../docs/format';
-import { InputEmulator, renderInputState, type InputCommand } from './utils/InputEmulator';
-
-const createExec = props => {
-  let getVal = null;
-  let execCommand = null;
-  let stateValue_ = null;
-
-  TestRenderer.create(
-    <Value initial={''}>
-      {input => {
-        stateValue_ = input.value;
-
-        return (
-          <Rifm value={input.value} onChange={input.set} {...props}>
-            {({ value, onChange }) => (
-              <InputEmulator value={value} onChange={onChange}>
-                {(exec, val) => {
-                  execCommand = exec;
-                  getVal = val;
-                  return null;
-                }}
-              </InputEmulator>
-            )}
-          </Rifm>
-        );
-      }}
-    </Value>
-  );
-
-  const exec = (cmd: InputCommand) => {
-    act(() => {
-      if (!execCommand) {
-        throw Error('rifm is not initialized');
-      }
-
-      execCommand(cmd);
-    });
-
-    if (!getVal) {
-      throw Error('rifm is not initialized');
-    }
-
-    expect(stateValue_).toEqual(getVal().value);
-
-    return expect(renderInputState(getVal()));
-  };
-
-  return exec;
-};
+import { createExec } from './utils/exec';
 
 test('format works', async () => {
   const exec = createExec({ format: numberFormat });
@@ -86,9 +33,9 @@ test('format works', async () => {
   exec({ type: 'PUT_SYMBOL', payload: 'x' }).toMatchInlineSnapshot(`"1|2,345"`);
 });
 
-test('format with custom refuse works', async () => {
+test('format with custom accept works', async () => {
   const exec = createExec({
-    refuse: /[^\d.]/gi,
+    accept: /[\d.]/gi,
     format: currencyFormat2,
   });
 
@@ -147,7 +94,7 @@ test('format with custom refuse works', async () => {
 
 test('format with fixed point delete backspace', async () => {
   const exec = createExec({
-    refuse: /[^\d.]/gi,
+    accept: /[\d.]/gi,
     format: currencyFormat,
   });
